Modify app file to refactor the class
